Created App + Basic Funcionality
Using NPM + Create React App, the App has been set up to the most basic funcionality, been able to fill the form and apear on the list of attendees.

For Switching Between the main page to the attendees page, react-router-dom has been used, keeping the Header of the Page between seccions.